Silence warning about pointer-truncation on 64bit archs.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileFoundation@4879 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/ETReflection.m
@@ -339,9 +339,11 @@
 	return [obj isMemberOfClass: [ETObjectMirror class]] &&
 		[obj representedObject] == _object;
 }
-- (unsigned int) hash
+- (NSUInteger) hash
 {
-	return (unsigned int) _object;
+	// NOTE: The cast prevents a compiler warning about pointer truncation on
+	// 64bit systems.
+	return (uintptr_t) _object;
 }
 - (id <ETClassMirror>) classMirror
 {
